Notes_Generator/requirements.txt
streamlit
google-genai
reportlab
python-docx
requests
